requirements.txt
dvc==2.43.1
fastapi==0.89.1
gradio==3.16.2
httpx==0.23.3
mlflow==2.1.1
numpy==1.23.5
Pillow==9.4.0
pytest==7.2.1
python-multipart==0.0.5
PyYAML==6.0
tensorflow==2.11.0
